Add async webpack.resolve support
Switch the compiler to async mode, to allow webpack to resolve the
included resources. The compilation is done in multiple passes until
all paths are resolved (most of the time just 1 extra pass).

Change the compiler export to a factory function so data can be passed
from the loader to the compiler. Because of this structure the mapcache
is not needed anymore.

Create a util function to correctly generate template IDs that are the
same between the loader and the compiler (previously they could differ).
They always use the resolved webpack path as source for the ID.

Skip dependencies in the token parsing that are not just a string. They
will be handled at runtime, and are the responsibility from the template
author.

Also export the template tokens on the template function, so they can be
used to register the template under a different name at runtime. This is
useful when dealing with dynamic templates where you want to use
`require.context` to bundle and register all your templates in your
application bootstrap.

Updated the test helper to make empty context available.

Added a lot of comments in the code.

re #26