Update `fakeModuleSystem` to fake webpack.resolve-like behavior
When not including `.twig` as extension, it will be added automatically.
This is to create a scenario where a file can be resolve by webpack, and
two different paths lead to the same resource.

Also add a check on disk to allow cases where a template include value
only exists at runtime as registered template.